Updated VS project file for Variant class, and fixed a small warning.
git-svn-id: svn://svn.code.sf.net/p/stella/code/trunk@2727 8b62c5a3-ac7e-4cc8-8f21-d9a121418aba

+++ src/gui/LauncherDialog.cxx
@@ -174,7 +174,7 @@ LauncherDialog::LauncherDialog(OSystem* osystem, DialogContainer* parent,
   mySelectedItem = 0;  // Highlight 'Rom Listing'
 
   // Create an options dialog, similar to the in-game one
-  myOptions = new OptionsDialog(osystem, parent, this, w * 0.8, h * 0.8, true);
+  myOptions = new OptionsDialog(osystem, parent, this, int(w * 0.8), int(h * 0.8), true);
 
   // Create a game list, which contains all the information about a ROM that
   // the launcher needs
